feat: 질답 세트를 삭제할 때 연관된 하이라이팅 존재 여부 검증 로직 추가

+++ backend/src/main/java/com/shyashyashya/refit/domain/qnaset/service/QnaSetService.java
@@ -1,5 +1,6 @@
 package com.shyashyashya.refit.domain.qnaset.service;
 
+import static com.shyashyashya.refit.global.exception.ErrorCode.QNA_DELETE_FAILED;
 import static com.shyashyashya.refit.global.exception.ErrorCode.QNA_SET_NOT_FOUND;
 
 import com.shyashyashya.refit.domain.industry.service.validator.IndustryValidator;
@@ -86,6 +87,11 @@ public class QnaSetService {
     public void deleteQnaSet(Long qnaSetId) {
         QnaSet qnaSet = getValidatedQnaSet(qnaSetId);
         interviewValidator.validateInterviewReviewStatus(qnaSet.getInterview(), InterviewReviewStatus.QNA_SET_DRAFT);
+
+        if (pdfHighlightingRepository.existsByQnaSet(qnaSet)) {
+            throw new CustomException(QNA_DELETE_FAILED);
+        }
+
         qnaSetRepository.delete(qnaSet);
     }
